ClockSettings: Use a Vector as the time zone model's container type
The time zones were stored as a static Span until commit 18afc41928, and
are now stored in a Vector. By continuing to tell the ItemListModel that
the container is a Span, we create a temporary Span in its constructor,
which the model tries to hold a constant reference to. Use the default
Vector container type now instead to prevent creating such temporaries.